+++ lib/moneys.dart
@@ -2,8 +2,8 @@ library moneys;
 
 export 'src/enumerations/currency.dart';
 export 'src/enumerations/expense_or_income.dart';
-export 'src/enumerations/renewal.dart';
 export 'src/enumerations/money_transaction_method.dart';
+export 'src/enumerations/renewal.dart';
 export 'src/models/budget.dart';
 export 'src/models/money.dart';
 export 'src/models/money_transaction.dart';